design: 삭제 확인 메세지 출력

+++ src/main/resources/static/js/eventAjax.js
@@ -90,6 +90,10 @@ function updateDetailViewData(eventId) {
 function deleteEvent() {
     var eventId = $('#eventId').val();
 
+    if (confirm("삭제하시겠습니까?") === false) {
+        return;
+    }
+
     $.ajax({
         url: "/event/" + eventId,
         type: "DELETE",
